requirements.txt
python-dotenv
requests
watchfiles
